+++ prettier.config.mjs
@@ -1,0 +1,7 @@
+export default {
+  printWidth: 100,
+  semi: true,
+  singleQuote: false,
+  tabWidth: 2,
+  trailingComma: "all",
+};